+++ e_metrobus/navigation/constants.py
@@ -1,18 +1,20 @@
-
 import math
+from collections import namedtuple
+from dataclasses import dataclass, fields
+
+from django.utils.translation import gettext as _
 
 ELLIPSE_RADIUS = 43
 ELLIPSE_X_OFFSET = 7
 ELLIPSE_Y_OFFSET = 7
 
+Consumption = namedtuple(
+    "Consumption", ["distance", "fuel", "co2", "nitrogen", "fine_dust"]
+)
 
-STATIONS = [
-    "Hertzallee",
-    "S+U Zoologischer Garten",
-    "S+U Zoologischer Garten/Jebensstr.",
-    "Breitscheidplatz",
-    "Budapester Str.",
-]
+FLEET_CONSUMPTION = Consumption(
+    distance=3000, fuel=200000, co2=300000, nitrogen=20000, fine_dust=10000
+)
 
 
 class Ellipse:
@@ -22,7 +24,73 @@ class Ellipse:
         angle = share * 2 * math.pi
         self.large = 0 if share <= 0.5 else 1
         self.x = ELLIPSE_X_OFFSET + ELLIPSE_RADIUS + ELLIPSE_RADIUS * math.sin(angle)
-        self.y = ELLIPSE_Y_OFFSET + ELLIPSE_RADIUS + ELLIPSE_RADIUS * math.cos(angle + math.pi)
+        self.y = (
+            ELLIPSE_Y_OFFSET
+            + ELLIPSE_RADIUS
+            + ELLIPSE_RADIUS * math.cos(angle + math.pi)
+        )
 
     def __str__(self):
         return f"M {ELLIPSE_RADIUS + ELLIPSE_X_OFFSET} {ELLIPSE_Y_OFFSET} A {ELLIPSE_RADIUS} {ELLIPSE_RADIUS} 0 {self.large} 1 {self.x} {self.y}"
+
+
+@dataclass
+class DataPerKilometer:
+    fuel: float
+    co2: float
+    nitrogen: float
+    fine_dust: float
+
+    def __mul__(self, value):
+        return DataPerKilometer(
+            *(getattr(self, dim.name) * value for dim in fields(self))
+        )
+
+    def __truediv__(self, value):
+        return DataPerKilometer(
+            *(getattr(self, dim.name) / value for dim in fields(self))
+        )
+
+
+@dataclass
+class Vehicle:
+    name: str
+    label: str
+    passengers: float
+    data: DataPerKilometer
+
+
+VEHICLES = [
+    Vehicle(
+        name="car",
+        label=_("PKW"),
+        passengers=1,
+        data=DataPerKilometer(fuel=0, co2=220.5, nitrogen=0.645, fine_dust=0.006),
+    ),
+    Vehicle(
+        name="bus",
+        label=_("Dieselbus"),
+        passengers=30,
+        data=DataPerKilometer(fuel=0, co2=1520, nitrogen=6.08, fine_dust=0.057),
+    ),
+    Vehicle(
+        name="e-bus",
+        label=_("Elektrobus"),
+        passengers=30,
+        data=DataPerKilometer(0, 0, 0, 0),
+    ),
+    Vehicle(
+        name="bicycle",
+        label=_("Fahrrad"),
+        passengers=30,
+        data=DataPerKilometer(0, 0, 0, 0),
+    ),
+    Vehicle(
+        name="pedestrian",
+        label=_("zu Fuß"),
+        passengers=30,
+        data=DataPerKilometer(0, 0, 0, 0),
+    ),
+]
+
+DATA_SOURCE = 'Umweltbundesamt, "Vergleich der durchschnittlicher Emissionen einzelner Verkehrsmittel im Personenverkehr in Deutschland - Bezugsjahr 2018", 01/2020'