+++ e_metrobus/navigation/widgets.py
@@ -1,10 +1,13 @@
 from collections import ChainMap
 
+from django.forms import widgets
+from django.forms.renderers import get_default_renderer
+from django.template.context_processors import csrf
+from django.templatetags.static import static
 from django.utils.html import format_html
 from django.utils.safestring import mark_safe
-from django.forms.renderers import get_default_renderer
-from django.templatetags.static import static
-from django.template.context_processors import csrf
+
+from e_metrobus.navigation import constants
 
 
 class CustomWidget:
@@ -60,7 +63,14 @@ class TopBarWidget(CustomWidget):
     template_name = "widgets/top_bar.html"
 
     def __init__(
-        self, title, title_icon, back_url, points, title_alt=None, template=None
+        self,
+        title,
+        title_icon,
+        back_url,
+        score,
+        title_alt=None,
+        template=None,
+        request=None,
     ):
         if template:
             self.template_name = template
@@ -68,7 +78,9 @@ class TopBarWidget(CustomWidget):
         self.title_icon = title_icon
         self.title_alt = title if title_alt is None else title_alt
         self.back_url = back_url
-        self.points = points
+        self.score = score
+        self.score_changed = False
+        self.request = request
 
 
 class FooterWidget(CustomWidget):
@@ -82,7 +94,7 @@ class FooterWidget(CustomWidget):
         },
         "leaf": {
             "name": "leaf",
-            "url": "navigation:dashboard",
+            "url": "navigation:environment",
             "enabled": False,
             "selected": False,
         },
@@ -110,3 +122,21 @@ class FooterWidget(CustomWidget):
 
     def get_context(self, **kwargs):
         return {"links": self.links}
+
+
+class FeedbackStarsWidget(widgets.NumberInput):
+    template_name = "widgets/feedback_stars.html"
+
+
+class FeedbackCommentWidget(widgets.TextInput):
+    template_name = "widgets/feedback_comment.html"
+
+
+class InfoTable(CustomWidget):
+    template_name = "widgets/info_table.html"
+
+    def get_context(self, **kwargs):
+        return {
+            "vehicles": constants.VEHICLES,
+            "source": constants.DATA_SOURCE
+        }